+++ CHANGELOG.md
@@ -5,9 +5,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 ## [Unreleased]
 
 ### Added
-- Support HEAD, CONNECT, and TRACE HTTP methods
+
+- Support HEAD, CONNECT, and TRACE HTTP methods.
+
+### Changed
+
+- Privatise `router.Route` struct method.
 
 ## [v0.1.0] - 2023-09-05
 
 ### Added
-- Add router implementation
+
+- Add router implementation.